CTCTRADERS-2836: Added packages CYA controller and view.
