+++ conf/views/addItems/packagesCheckYourAnswers.njk
@@ -1,0 +1,45 @@
+{% extends "includes/layout.njk" %}
+
+{% from "govuk/components/back-link/macro.njk"     import govukBackLink %}
+{% from "macros/title.njk"                         import title %}
+{% from "macros/section.njk"                       import section as sectionMacro %}
+{% from "govuk/components/button/macro.njk"        import govukButton %}
+
+{% block pageTitle %}
+  {{ title(messages("packagesCheckYourAnswers.title")) }}
+{% endblock %}
+
+{% block mainContent %}
+
+  {{ govukBackLink({
+    text: messages("site.back")
+  }) }}
+
+  <div class="govuk-main-wrapper">
+    <div class="govuk-grid-row">
+      <div class="govuk-grid-column-two-thirds">
+
+        <span class="govuk-caption-xl">{{ messages("packages.preHeading") }}</span>
+
+        <h1 class="govuk-heading-xl">
+          {{ messages("packagesCheckYourAnswers.heading") }}
+        </h1>
+
+        {% for section in sections %}
+          {{ sectionMacro(section.sectionTitle, section.rows, section.addAnother, section.sectionSubTitle) }}
+        {% endfor %}
+
+        {{ csrf() | safe }}
+
+        {{ govukButton({
+          text: messages("site.continue"),
+          preventDoubleClick: true,
+          href: nextPageUrl,
+          attributes: {id: "submit"}
+        }) }}
+
+      </div>
+    </div>
+  </div>
+
+{% endblock %}
