+++ test/controllers/addItems/packagesInformation/AddAnotherPackageControllerSpec.scala
@@ -14,19 +14,20 @@
  * limitations under the License.
  */
 
-package controllers.addItems
+package controllers.addItems.packagesInformation
 
 import base.{MockNunjucksRendererApp, SpecBase}
 import controllers.{routes => mainRoutes}
-import forms.addItems.RemovePackageFormProvider
+import forms.addItems.AddAnotherPackageFormProvider
 import matchers.JsonMatchers
-import models.NormalMode
+import models.{NormalMode, UserAnswers}
 import navigation.annotations.addItems.AddItemsPackagesInfo
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentCaptor
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{times, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
+import pages.addItems.AddAnotherPackagePage
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.json.{JsObject, Json}
@@ -38,29 +39,28 @@ import uk.gov.hmrc.viewmodels.{NunjucksSupport, Radios}
 
 import scala.concurrent.Future
 
-class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
+class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
 
   def onwardRoute = Call("GET", "/foo")
 
-  private val formProvider = new RemovePackageFormProvider()
-  private val form         = formProvider(index.display)
-  private val template     = "addItems/removePackage.njk"
-
-  lazy val removePackageRoute = controllers.addItems.packagesInformation.routes.RemovePackageController.onPageLoad(lrn, index, index, NormalMode).url
+  val formProvider = new AddAnotherPackageFormProvider()
+  val form         = formProvider()
 
   override def guiceApplicationBuilder(): GuiceApplicationBuilder =
     super
       .guiceApplicationBuilder()
       .overrides(bind(classOf[Navigator]).qualifiedWith(classOf[AddItemsPackagesInfo]).toInstance(new FakeNavigator(onwardRoute)))
 
-  "RemovePackage Controller" - {
+  lazy val addAnotherPackageRoute = controllers.addItems.packagesInformation.routes.AddAnotherPackageController.onPageLoad(lrn, index, NormalMode).url
+
+  "AddAnotherPackage Controller" - {
 
     "must return OK and the correct view for a GET" in {
       dataRetrievalWithData(emptyUserAnswers)
       when(mockRenderer.render(any(), any())(any()))
         .thenReturn(Future.successful(Html("")))
 
-      val request        = FakeRequest(GET, removePackageRoute)
+      val request        = FakeRequest(GET, addAnotherPackageRoute)
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
 
@@ -71,27 +71,56 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form"         -> form,
-        "itemIndex"    -> itemIndex.display,
-        "packageIndex" -> packageIndex.display,
-        "mode"         -> NormalMode,
-        "lrn"          -> lrn,
-        "radios"       -> Radios.yesNo(form("value"))
+        "form"      -> form,
+        "mode"      -> NormalMode,
+        "lrn"       -> lrn,
+        "itemIndex" -> itemIndex.display,
+        "radios"    -> Radios.yesNo(form("value"))
       )
 
-      val jsonWithoutConfig = jsonCaptor.getValue - configKey
+      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
+      jsonCaptor.getValue must containJson(expectedJson)
 
-      templateCaptor.getValue mustEqual template
-      jsonWithoutConfig mustBe expectedJson
+    }
+
+    "must populate the view correctly on a GET when the question has previously been answered" in {
+
+      val userAnswers = UserAnswers(lrn, eoriNumber).set(AddAnotherPackagePage(index), true).success.value
+      dataRetrievalWithData(userAnswers)
+      when(mockRenderer.render(any(), any())(any()))
+        .thenReturn(Future.successful(Html("")))
+
+      val request        = FakeRequest(GET, addAnotherPackageRoute)
+      val templateCaptor = ArgumentCaptor.forClass(classOf[String])
+      val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
+
+      val result = route(app, request).value
+
+      status(result) mustEqual OK
+
+      verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
+
+      val filledForm = form.bind(Map("value" -> "true"))
+
+      val expectedJson = Json.obj(
+        "form"      -> filledForm,
+        "mode"      -> NormalMode,
+        "lrn"       -> lrn,
+        "itemIndex" -> itemIndex.display,
+        "radios"    -> Radios.yesNo(filledForm("value"))
+      )
+
+      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
+      jsonCaptor.getValue must containJson(expectedJson)
+
     }
 
     "must redirect to the next page when valid data is submitted" in {
       dataRetrievalWithData(emptyUserAnswers)
-
       when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
       val request =
-        FakeRequest(POST, removePackageRoute)
+        FakeRequest(POST, addAnotherPackageRoute)
           .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
@@ -99,6 +128,7 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual onwardRoute.url
+
     }
 
     "must return a Bad Request and errors when invalid data is submitted" in {
@@ -106,7 +136,7 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       when(mockRenderer.render(any(), any())(any()))
         .thenReturn(Future.successful(Html("")))
 
-      val request        = FakeRequest(POST, removePackageRoute).withFormUrlEncodedBody(("value", ""))
+      val request        = FakeRequest(POST, addAnotherPackageRoute).withFormUrlEncodedBody(("value", ""))
       val boundForm      = form.bind(Map("value" -> ""))
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
@@ -118,31 +148,29 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form"         -> boundForm,
-        "itemIndex"    -> itemIndex.display,
-        "packageIndex" -> packageIndex.display,
-        "mode"         -> NormalMode,
-        "lrn"          -> lrn,
-        "radios"       -> Radios.yesNo(boundForm("value"))
+        "form"      -> boundForm,
+        "mode"      -> NormalMode,
+        "lrn"       -> lrn,
+        "itemIndex" -> itemIndex.display,
+        "radios"    -> Radios.yesNo(boundForm("value"))
       )
 
-      val jsonWithoutConfig = jsonCaptor.getValue - configKey
+      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
+      jsonCaptor.getValue must containJson(expectedJson)
 
-      templateCaptor.getValue mustEqual template
-      jsonWithoutConfig mustBe expectedJson
     }
 
     "must redirect to Session Expired for a GET if no existing data is found" in {
 
       dataRetrievalNoData()
-
-      val request = FakeRequest(GET, removePackageRoute)
+      val request = FakeRequest(GET, addAnotherPackageRoute)
 
       val result = route(app, request).value
 
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual mainRoutes.SessionExpiredController.onPageLoad().url
+
     }
 
     "must redirect to Session Expired for a POST if no existing data is found" in {
@@ -150,7 +178,7 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       dataRetrievalNoData()
 
       val request =
-        FakeRequest(POST, removePackageRoute)
+        FakeRequest(POST, addAnotherPackageRoute)
           .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
@@ -158,6 +186,7 @@ class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp 
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual mainRoutes.SessionExpiredController.onPageLoad().url
+
     }
   }
 }
