+++ test/controllers/addItems/packagesInformation/RemovePackageControllerSpec.scala
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package controllers.addItems
+package controllers.addItems.packagesInformation
 
 import base.{MockNunjucksRendererApp, SpecBase}
 import controllers.{routes => mainRoutes}
-import forms.addItems.AddAnotherPackageFormProvider
+import forms.addItems.RemovePackageFormProvider
 import matchers.JsonMatchers
-import models.{NormalMode, UserAnswers}
+import models.NormalMode
 import navigation.annotations.addItems.AddItemsPackagesInfo
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentCaptor
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{times, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
-import pages.addItems.AddAnotherPackagePage
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.json.{JsObject, Json}
@@ -39,28 +38,29 @@ import uk.gov.hmrc.viewmodels.{NunjucksSupport, Radios}
 
 import scala.concurrent.Future
 
-class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
+class RemovePackageControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
 
   def onwardRoute = Call("GET", "/foo")
 
-  val formProvider = new AddAnotherPackageFormProvider()
-  val form         = formProvider()
+  private val formProvider = new RemovePackageFormProvider()
+  private val form         = formProvider(index.display)
+  private val template     = "addItems/removePackage.njk"
+
+  lazy val removePackageRoute = controllers.addItems.packagesInformation.routes.RemovePackageController.onPageLoad(lrn, index, index, NormalMode).url
 
   override def guiceApplicationBuilder(): GuiceApplicationBuilder =
     super
       .guiceApplicationBuilder()
       .overrides(bind(classOf[Navigator]).qualifiedWith(classOf[AddItemsPackagesInfo]).toInstance(new FakeNavigator(onwardRoute)))
 
-  lazy val addAnotherPackageRoute = controllers.addItems.packagesInformation.routes.AddAnotherPackageController.onPageLoad(lrn, index, NormalMode).url
-
-  "AddAnotherPackage Controller" - {
+  "RemovePackage Controller" - {
 
     "must return OK and the correct view for a GET" in {
       dataRetrievalWithData(emptyUserAnswers)
       when(mockRenderer.render(any(), any())(any()))
         .thenReturn(Future.successful(Html("")))
 
-      val request        = FakeRequest(GET, addAnotherPackageRoute)
+      val request        = FakeRequest(GET, removePackageRoute)
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
 
@@ -71,56 +71,27 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form"      -> form,
-        "mode"      -> NormalMode,
-        "lrn"       -> lrn,
-        "itemIndex" -> itemIndex.display,
-        "radios"    -> Radios.yesNo(form("value"))
+        "form"         -> form,
+        "itemIndex"    -> itemIndex.display,
+        "packageIndex" -> packageIndex.display,
+        "mode"         -> NormalMode,
+        "lrn"          -> lrn,
+        "radios"       -> Radios.yesNo(form("value"))
       )
 
-      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
-      jsonCaptor.getValue must containJson(expectedJson)
+      val jsonWithoutConfig = jsonCaptor.getValue - configKey
 
-    }
-
-    "must populate the view correctly on a GET when the question has previously been answered" in {
-
-      val userAnswers = UserAnswers(lrn, eoriNumber).set(AddAnotherPackagePage(index), true).success.value
-      dataRetrievalWithData(userAnswers)
-      when(mockRenderer.render(any(), any())(any()))
-        .thenReturn(Future.successful(Html("")))
-
-      val request        = FakeRequest(GET, addAnotherPackageRoute)
-      val templateCaptor = ArgumentCaptor.forClass(classOf[String])
-      val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
-
-      val result = route(app, request).value
-
-      status(result) mustEqual OK
-
-      verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
-
-      val filledForm = form.bind(Map("value" -> "true"))
-
-      val expectedJson = Json.obj(
-        "form"      -> filledForm,
-        "mode"      -> NormalMode,
-        "lrn"       -> lrn,
-        "itemIndex" -> itemIndex.display,
-        "radios"    -> Radios.yesNo(filledForm("value"))
-      )
-
-      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
-      jsonCaptor.getValue must containJson(expectedJson)
-
+      templateCaptor.getValue mustEqual template
+      jsonWithoutConfig mustBe expectedJson
     }
 
     "must redirect to the next page when valid data is submitted" in {
       dataRetrievalWithData(emptyUserAnswers)
+
       when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
       val request =
-        FakeRequest(POST, addAnotherPackageRoute)
+        FakeRequest(POST, removePackageRoute)
           .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
@@ -128,7 +99,6 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual onwardRoute.url
-
     }
 
     "must return a Bad Request and errors when invalid data is submitted" in {
@@ -136,7 +106,7 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       when(mockRenderer.render(any(), any())(any()))
         .thenReturn(Future.successful(Html("")))
 
-      val request        = FakeRequest(POST, addAnotherPackageRoute).withFormUrlEncodedBody(("value", ""))
+      val request        = FakeRequest(POST, removePackageRoute).withFormUrlEncodedBody(("value", ""))
       val boundForm      = form.bind(Map("value" -> ""))
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
@@ -148,29 +118,31 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form"      -> boundForm,
-        "mode"      -> NormalMode,
-        "lrn"       -> lrn,
-        "itemIndex" -> itemIndex.display,
-        "radios"    -> Radios.yesNo(boundForm("value"))
+        "form"         -> boundForm,
+        "itemIndex"    -> itemIndex.display,
+        "packageIndex" -> packageIndex.display,
+        "mode"         -> NormalMode,
+        "lrn"          -> lrn,
+        "radios"       -> Radios.yesNo(boundForm("value"))
       )
 
-      templateCaptor.getValue mustEqual "addItems/addAnotherPackage.njk"
-      jsonCaptor.getValue must containJson(expectedJson)
+      val jsonWithoutConfig = jsonCaptor.getValue - configKey
 
+      templateCaptor.getValue mustEqual template
+      jsonWithoutConfig mustBe expectedJson
     }
 
     "must redirect to Session Expired for a GET if no existing data is found" in {
 
       dataRetrievalNoData()
-      val request = FakeRequest(GET, addAnotherPackageRoute)
+
+      val request = FakeRequest(GET, removePackageRoute)
 
       val result = route(app, request).value
 
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual mainRoutes.SessionExpiredController.onPageLoad().url
-
     }
 
     "must redirect to Session Expired for a POST if no existing data is found" in {
@@ -178,7 +150,7 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       dataRetrievalNoData()
 
       val request =
-        FakeRequest(POST, addAnotherPackageRoute)
+        FakeRequest(POST, removePackageRoute)
           .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
@@ -186,7 +158,6 @@ class AddAnotherPackageControllerSpec extends SpecBase with MockNunjucksRenderer
       status(result) mustEqual SEE_OTHER
 
       redirectLocation(result).value mustEqual mainRoutes.SessionExpiredController.onPageLoad().url
-
     }
   }
 }
